// src/domain/db/enum/db.enums.ts
export enum JobType {
  Scheduled = 'scheduled',
  Manual = 'manual',
}

export enum JobStatus {
  Running = 'running',
  Completed = 'completed',
  Failed = 'failed',
}

export enum PageStatus {
  Pending = 'pending',
  Success = 'success',
  Failed = 'failed',
}

export enum ImageAssetStatus {
  Pending = 'pending',
  Uploaded = 'uploaded',
  Failed = 'failed',
}